Refactor daemon process supervision modules
